move bubbletea program to spec.go
